+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### New
 
 - New field `retry_as_batch` added to the `kafka` output to assist in ensuring message ordering through retries.
+- Field `delay_period` added to the experimental `aws_s3` input.
 
 ### Fixed